backend/requirements.txt
fastapi
uvicorn
pydantic
pypdf
tiktoken
faiss-cpu
numpy
openai
reportlab
streamlit
requests
python-dotenv
python-docx
openpyxl
python-multipart
